FDTD | Cleanup | Separate pyWrapper unit tests
