+++ test/pyWrapper/conftest.py
@@ -3,9 +3,15 @@
 import os
 import re
 import shutil
+import sys
 from pathlib import Path
 
 import pytest
+
+
+PROJECT_ROOT = Path(__file__).resolve().parents[2]
+if str(PROJECT_ROOT) not in sys.path:
+    sys.path.insert(0, str(PROJECT_ROOT))
 
 
 def pytest_addoption(parser):
